# backend/static/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Produucts": {
				"properties": {
					"categoryId": {
						"type": "string"
					},
					"discontinued": {
						"type": "boolean"
					},
					"productName": {
						"type": "string"
					},
					"quantityPerUnit": {
						"type": "string"
					},
					"supplierId": {
						"type": "string"
					},
					"unitPrice": {
						"type": "string"
					},
					"unitsInStock": {
						"type": "number",
						"format": "double"
					},
					"unitsOnOrder": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"categoryId",
					"discontinued",
					"productName",
					"quantityPerUnit",
					"supplierId",
					"unitPrice",
					"unitsInStock",
					"unitsOnOrder"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"keycloak": {
				"type": "apiKey",
				"name": "Authorization",
				"description": "Keycloak Bearer Token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/Product": {
			"get": {
				"operationId": "GetProductAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProduuct",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"UnitsOnOrder": {
													"type": "number",
													"format": "double"
												},
												"UnitsInStock": {
													"type": "number",
													"format": "double"
												},
												"UnitPrice": {
													"type": "string"
												},
												"SupplierId": {
													"type": "string"
												},
												"QuantityPerUnit": {
													"type": "string"
												},
												"ProductName": {
													"type": "string"
												},
												"Discontinued": {
													"type": "boolean"
												},
												"CategoryId": {
													"type": "string"
												},
												"Id": {
													"type": "string"
												}
											},
											"required": [
												"UnitsOnOrder",
												"UnitsInStock",
												"UnitPrice",
												"SupplierId",
												"QuantityPerUnit",
												"ProductName",
												"Discontinued",
												"CategoryId",
												"Id"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"ผู้ใช้งานนี้ไม่สามารถเพิ่มข้อมูลได้",
												"ไม่่พบข้อมุลประเภท"
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Produucts"
							}
						}
					}
				}
			}
		},
		"/Product/{id}": {
			"get": {
				"operationId": "GetProductById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdataProduuct",
				"responses": {
					"200": {
						"description": "Update",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Produucts"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Delete",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}